backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
marshmallow==3.20.1
